Add files via upload
FinalTR.ipynb is used to create a model where the extracted column, Good_Rate is not considered for taking inputs from the user. The functionality is provided for the customers who want to check their eligibility for credit card approval but do not have any previous banking record to fill (i.e. The Rate value cannot be decided in their case)